Rewrite the pipeline to remove a lot of duplication
I'm not sure why we have logic to check if it's not a PR run or an internal run if the pipeline is for PRs only. Maybe this won't work, and I don't understand the system that well, but it seems like it may have been copy pasted and a result of de-merging two pipelines into separate ones. There is also a large amount of yaml which is duplicated and can be simplified/shared.